+++ changes.md
@@ -1,7 +1,7 @@
 ## v1.2.0 (2026-xxx)
 
 - Adds NIRDA and VISDA classes which contain accurate and up to date parameters to perform timing and data volume calculations.
-- Adds overhead class which accounts for pre- and post- overhead timings for both VISDA and NIRDA
+- Adds overhead class which accounts for pre- and post- overhead timings for both VISDA and NIRDA.
 - Adds baseline short-term calendar runner script to docs/
 - Adds ability to merge back-to-back observations of the same target.
 - Adds ability to override payload parameters set by original long-term calendar on a per-priority type basis.
@@ -11,4 +11,6 @@
 - Adds log file to track changes, info, and warnings raised by the short term scheduler.
 - Fixes minute-by-minute parsing to improve processing time.
 - Adds several progress bars during various slower processing sections.
+- Adds helper method to generate diagnostic data file.
+  - Diag file contains a observation file manifest including compressed fits file names.
 - Adds tests for all of these changes.